simplify qr1 driver
